requirements.txt
numpy==1.16.2
tensorflow==2.0.0
scipy==1.4.1
pandas==1.0.3
gensim==3.8.1
nltk==3.4.5
PyYAML==5.3.1
